requirements.txt
bcrypt==3.1.7
boto3==1.11.13
botocore==1.14.13
certifi==2019.11.28
cffi==1.13.2
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Script==2.0.6
flask-swagger-ui==3.20.9
itsdangerous==1.1.0
Jinja2==2.11.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.19
Pillow==7.0.0
protobuf==3.6.1
pycparser==2.19
PyJWT==1.7.1
six==1.14.0
Werkzeug==0.16.1
smart-open==1.9.0
Flask-API==2.0
